third_party/WebKit/LayoutTests/fast/text-autosizing/style-recalc-no-layout.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
<script src="resources/autosizingTest.js"></script>
<style>
#autosize {
  font-size: 10px;
}
</style>
</head>
<body>
  <div id="autosize">
    This test verifies that a style recalc on a block with an autosizing multiplier does not
    cause an unnecessary relayout.
    Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
  </div>
  <script>
    // Force style recalc and layout.
    document.body.offsetTop;

    if (window.internals)
        shouldBe("internals.needsLayoutCount()", "0", true);

    // Should cause a repaint but no layout.
    document.querySelector('#autosize').style.color = 'green';

    if (window.internals) {
        shouldBe("internals.updateStyleAndReturnAffectedElementCount()", "1", true);
        shouldBe("internals.needsLayoutCount()", "0", true);
    }
  </script>
</body>
</html>
